+++ packages/stream_video/CHANGELOG.md
@@ -1,5 +1,8 @@
-## Unregistered
+## Unreleased
 
+🐞 Fixed
+* Handled SFU stats reporting failures gracefully
+* 
 ✅ Added
 * Added option to configure android audio configuration when initializing `StreamVideo` instance by providing `androidAudioConfiguration` to `StreamVideoOptions`.