+++ packages/stream_video/CHANGELOG.md
@@ -1,8 +1,8 @@
-## Unreleased
+## 0.10.3
 
 🐞 Fixed
 * Handled SFU stats reporting failures gracefully
-* 
+
 ✅ Added
 * Added option to configure android audio configuration when initializing `StreamVideo` instance by providing `androidAudioConfiguration` to `StreamVideoOptions`.